Add endpoints p/ criar relacionamento entre Endereço - Usuário/Produto
