+++ app/Http/Resources/AddressUserProductResource.php
@@ -5,7 +5,7 @@ namespace App\Http\Resources;
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
 
-class AdressUserProductResource extends JsonResource
+class AddressUserProductResource extends JsonResource
 {
     /**
      * Transform the resource into an array.
